ADIF getting wrong freq when no rig control; also microwaves freqs were wrong

+++ mqt/MqtLogger/LoggerContacts.cpp
@@ -532,7 +532,7 @@ QString ContestContact::getADIFLine()
 
 
     QString cb;
-    long txfreq = contest->getAdifFreqBand(frequency.getValue(), cb);
+    double txfreq = contest->getAdifFreqBand(frequency.getValue(), cb);
     outstr += makeADIFField( "BAND", cb );
 
     double dfreq = txfreq/1000000.0;  // MHz
